Move requests to dashboard service #530
